% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsetup.R
\name{create_qc_status_md}
\alias{create_qc_status_md}
\title{Create project QC_STATUS.md}
\usage{
create_qc_status_md(path, overwrite = FALSE)
}
\arguments{
\item{path}{Project directory.}

\item{overwrite}{Logical. If \code{TRUE}, overwrite an existing \code{QC_STATUS.md}.}
}
\value{
Invisibly returns the \code{QC_STATUS.md} path.
}
\description{
Create project QC_STATUS.md
}
